Add files via upload
This is the 5th challenge of the first set. The previous 4 were lost to the passage of time and the recycle bin but are solvable in even less lines of code than this one. When run it encrypts a plaintext with a given key by XOR'ing their respective bytes in sequential order.